# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Audiobookshelf API",
    "version": "0.1.0",
    "description": "Audiobookshelf API with autogenerated OpenAPI doc"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/podcasts": {
      "post": {
        "summary": "Create a new podcast",
        "operationId": "createPodcast",
        "tags": [
          "Podcasts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Podcast"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a podcast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Podcast"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/podcasts/feed": {
      "post": {
        "summary": "Get podcast feed",
        "operationId": "getPodcastFeed",
        "tags": [
          "Podcasts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rssFeed": {
                    "type": "string",
                    "description": "The RSS feed URL of the podcast"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved podcast feed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "podcast": {
                      "$ref": "#/components/schemas/Podcast"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/podcasts/opml/parse": {
      "post": {
        "summary": "Get feeds from OPML text",
        "description": "Parse OPML text and return an array of feeds",
        "operationId": "getFeedsFromOPMLText",
        "tags": [
          "Podcasts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "opmlText": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully parsed OPML text and returned feeds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "feeds": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          },
                          "feedUrl": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, OPML text not provided"
          },
          "403": {
            "description": "Forbidden, user is not admin"
          }
        }
      }
    },
    "/api/podcasts/opml/create": {
      "post": {
        "summary": "Bulk create podcasts from OPML feed URLs",
        "operationId": "bulkCreatePodcastsFromOpmlFeedUrls",
        "tags": [
          "Podcasts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "feeds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "libraryId": {
                    "$ref": "#/components/schemas/libraryId"
                  },
                  "folderId": {
                    "$ref": "#/components/schemas/folderId"
                  },
                  "autoDownloadEpisodes": {
                    "$ref": "#/components/schemas/autoDownloadEpisodes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created podcasts from feed URLs"
          },
          "400": {
            "description": "Bad request, invalid request body"
          },
          "403": {
            "description": "Forbidden, user is not admin"
          },
          "404": {
            "description": "Folder not found"
          }
        }
      }
    },
    "/api/podcasts/{id}/checknew": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "get": {
        "summary": "Check and download new episodes",
        "operationId": "checkNewEpisodes",
        "tags": [
          "Podcasts"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of episodes to download",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully checked and downloaded new episodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "episodes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PodcastEpisode"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/podcasts/{id}/clear-queue": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "get": {
        "summary": "Clear episode download queue",
        "operationId": "clearEpisodeDownloadQueue",
        "tags": [
          "Podcasts"
        ],
        "responses": {
          "200": {
            "description": "Successfully cleared download queue"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/podcasts/{id}/downloads": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "get": {
        "summary": "Get episode downloads",
        "operationId": "getEpisodeDownloads",
        "tags": [
          "Podcasts"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved episode downloads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "downloads": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PodcastEpisode"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/podcasts/{id}/search-episode": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "get": {
        "summary": "Find episode by title",
        "operationId": "findEpisode",
        "tags": [
          "Podcasts"
        ],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Title of the episode to search for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found episodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "episodes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PodcastEpisode"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/podcasts/{id}/download-episodes": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "post": {
        "summary": "Download podcast episodes",
        "operationId": "downloadEpisodes",
        "tags": [
          "Podcasts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully started episode download"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/podcasts/{id}/match-episodes": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "post": {
        "summary": "Quick match podcast episodes",
        "operationId": "quickMatchEpisodes",
        "tags": [
          "Podcasts"
        ],
        "parameters": [
          {
            "name": "override",
            "in": "query",
            "description": "Override existing details if set to 1",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully matched episodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "numEpisodesUpdated": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/podcasts/{id}/episode/{episodeId}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Podcast ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        },
        {
          "name": "episodeId",
          "in": "path",
          "description": "Episode ID",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/podcastId"
          }
        }
      ],
      "patch": {
        "summary": "Update a podcast episode",
        "operationId": "updateEpisode",
        "tags": [
          "Podcasts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated episode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Podcast"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "get": {
        "summary": "Get a specific podcast episode",
        "operationId": "getEpisode",
        "tags": [
          "Podcasts"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved episode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastEpisode"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Remove a podcast episode",
        "operationId": "removeEpisode",
        "tags": [
          "Podcasts"
        ],
        "parameters": [
          {
            "name": "hard",
            "in": "query",
            "description": "Hard delete the episode if set to 1",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully removed episode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Podcast"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer authentication",
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "libraryItemId": {
        "type": "string",
        "description": "The ID of library items after 2.3.0.",
        "format": "uuid",
        "example": "e4bb1afb-4a4f-4dd6-8be0-e615d233185b"
      },
...
      "libraryId": {
        "type": "string",
        "description": "The ID of the library.",
        "format": "uuid",
        "example": "e4bb1afb-4a4f-4dd6-8be0-e615d233185b"
      },
      "folderId": {
        "type": "string",
        "description": "The ID of the folder.",
        "format": "uuid",
        "example": "e4bb1afb-4a4f-4dd6-8be0-e615d233185b"
      },
...
      "podcastId": {
        "type": "string",
        "description": "The ID of podcasts and podcast episodes after 2.3.0.",
        "format": "uuid",
        "example": "e4bb1afb-4a4f-4dd6-8be0-e615d233185b"
      },
...
      "autoDownloadEpisodes": {
        "type": "boolean",
        "description": "Whether episodes are automatically downloaded."
      },
      "Podcast": {
        "type": "object",
        "description": "A podcast containing multiple episodes.",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/podcastId"
          },
          "libraryItemId": {
            "$ref": "#/components/schemas/libraryItemId"
          },
          "metadata": {
            "$ref": "#/components/schemas/PodcastMetadata"
          },
          "coverPath": {
            "type": "string",
            "description": "The file path to the podcast's cover image.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "description": "The tags associated with the podcast.",
            "items": {
              "type": "string"
            }
          },
          "episodes": {
            "type": "array",
            "description": "The episodes of the podcast.",
            "items": {
              "$ref": "#/components/schemas/PodcastEpisode"
            }
          },
          "autoDownloadEpisodes": {
            "$ref": "#/components/schemas/autoDownloadEpisodes"
          },
          "autoDownloadSchedule": {
            "type": "string",
            "description": "The schedule for automatic episode downloads, in cron format.",
            "nullable": true
          },
          "lastEpisodeCheck": {
            "type": "integer",
            "description": "The timestamp of the last episode check."
          },
          "maxEpisodesToKeep": {
            "type": "integer",
            "description": "The maximum number of episodes to keep."
          },
          "maxNewEpisodesToDownload": {
            "type": "integer",
            "description": "The maximum number of new episodes to download when automatically downloading epsiodes."
          },
          "lastCoverSearch": {
            "type": "integer",
            "description": "The timestamp of the last cover search.",
            "nullable": true
          },
          "lastCoverSearchQuery": {
            "type": "string",
            "description": "The query used for the last cover search.",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "description": "The total size of all episodes in bytes."
          },
          "duration": {
            "type": "integer",
            "description": "The total duration of all episodes in seconds."
          },
          "numTracks": {
            "type": "integer",
            "description": "The number of tracks (episodes) in the podcast."
          },
          "latestEpisodePublished": {
            "type": "integer",
            "description": "The timestamp of the most recently published episode."
          }
        }
      }
    },
...
  }
}